third_party/WebKit/LayoutTests/http/tests/sendbeacon/beacon-blob-with-non-simple-type.html

<!DOCTYPE HTML>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws(
      'SecurityError',
      () => {
        navigator.sendBeacon("/",
                             new Blob(["X"], {type: "image/png"}));
      });
}, "navigator.sendBeacon() to a cross-origin target with a Blob body with " +
    "non-simple type should throw.");
</script>
